+++ OneBlink.SDK/models/ApprovalsFormsApp.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace OneBlink.SDK.Model
 {
     public class ApprovalsFormsApp : FormsAppBase
@@ -13,6 +15,10 @@ namespace OneBlink.SDK.Model
         public override string type
         {
             get;
+        }
+        public List<string> approvalGroups
+        {
+            get; set;
         }
     }
 }